refactor(observability): extract telemetry and audit loggers module (#188)
